requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
cryptography>=42.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
redis>=5.0.0
openai>=1.54.0
